# docs/openapi-spec.yaml
openapi: 3.0.0
info:
  title: Singularity Channels API
  description: API for interacting with Singularity channels
  version: 1.0.0
paths:
  /open:
    post:
      summary: Open a new channel
      description: |
       Example `curl` command:
        ```bash
        curl -X POST http://localhost:3001/open \
          -H "Content-Type: application/json" \
          -d '{
            "senderAddress": "addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8",
            "signerPubKey": "827FF8BEEEA147E7FA287DC5271E84049A0B6724BF47F66BC3748E469B210142",
            "receiverAddress": "addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8",
            "initialDeposit": 10000000000,
            "expirationDate": 1715800000000000,
            "groupId": "group-42"
          }'
        ```
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenChannelParams'
            example:
              senderAddress: "addr_test1vzwlx094fypzcfsz53n5gufc6h84tawp9pdyklm05pr385gl956ej"
              signerPubKey: "145562a792f2ca28ae87f1f7afd25b126f76547103204f62432999bcc2b5c542"
              receiverAddress: "addr_test1xyzvwx123fypzcfsz53n5gufc6h84tawp9pdyklm05pr385g"
              initialDeposit: "1000000"
              expirationDate: "1716148967"
              groupId: "ZC63QytbzJt75kHGkZ8PEWXXDl2GDFRVwUT1CI2lbwY="
      responses:
        "200":
          description: Channel opened successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  openChannelCbor:
                    type: string
                  channelId:
                    type: string
              example:
                openChannelCbor: "84a400d9..."
                channelId: "b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0"
  /update:
    post:
      summary: Update an existing channel by adding funds, extending the expiration date, or both.
      description: |
        Example `curl` command:

        ```bash
        curl -X POST http://localhost:3000/update \
          -H "Content-Type: application/json" \
          -d '{
            "channelId": "b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0",
            "userAddress": "addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8",
            "senderAddress": "addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8",
            "addDeposit": 5,
            "expirationDate": 1716158967
          }'

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateChannelParams'
            example:
              channelId: "b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0"
              addDeposit: "500000"
              expirationDate: "1716148967"
              userAddress: "addr_test1vzwlx094fypzcfsz53n5gufc6h84tawp9pdyklm05pr385gl956ej"
              senderAddress: "addr_test1vzwlx094fypzcfsz53n5gufc6h84tawp9pdyklm05pr385gl956ej"
      responses:
        "200":
          description: Channel updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  updateChannelCbor:
                    type: string
              example:
                updateChannelCbor: "84a400d9..."
  /claim:
    post:
      summary: Claim funds from a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ClaimParams'
            example:
              - channelId: "b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0"
                signature: "4eb27d97d08c53f4d5d53aabd395e7eaa3a095ac084db3201974146804e1bde1"
                amount: "1500000"
                finalize: false
                senderAddress: "addr_test1vzwlx094fypzcfsz53n5gufc6h84tawp9pdyklm05pr385gl956ej"
      responses:
        "200":
          description: Channel claim successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  claimChannelCbor:
                    type: string
              example:
                claimChannelCbor: "84a400d9..."
  /close:
    post:
      summary: Close an existing channel
      description: |
        Example `curl` command:

        ```bash
        curl -X POST http://localhost:3000/close \
          -H "Content-Type: application/json" \
          -d '{
            "senderAddress": "addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8",
            "channelId": "b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0"
          }'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimParams'
            example:
              channelId: "b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0"
              senderAddress: "addr_test1vzwlx094fypzcfsz53n5gufc6h84tawp9pdyklm05pr385gl956ej"
      responses:
        "200":
          description: Channel closed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  closedChannelCbor:
                    type: string
              example:
                closedChannelCbor: "84a400d9..."
  /channels:
    get:
      summary: Get all channels
      description: |
        Example `curl` command:
        ```bash
        curl http://localhost:3000/channels
      responses:
        "200":
          description: List of all channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
  /channel-with-id:
    get:
      summary: Get a channel by channelId
      description: |
        Example `curl` command:

        ```bash
        curl http://localhost:3000/channels-with-id\?channelId\="b126f76547103204f62432999bcc2b5cb126f76547103204f62432999bcc2b5c0"
      parameters:
        - name: channelId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
  /channels-from-sender:
    get:
      summary: Get channels by sender
      description: |
        Example `curl` command:

        ```bash
        curl http://localhost:3000/channels-from-sender\?senderAddress\=addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8
      parameters:
        - name: senderAddress
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of channels from sender
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
  /channels-from-receiver:
    get:
      summary: Get channels by receiver
      description: |
        Example `curl` command:

        ```bash
        curl http://localhost:3000/channels-from-receiver\?receiverAddress\=addr_test1qz6fcx7lay9czf838z4m0ppkyzzgspp9kz5uwede3ry4fmxmkhwgrz8c7pyr8dmqvcmf7sa2dagvvtfmw5585w847jgqwrhsv8
      parameters:
        - name: receiverAddress
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of channels from receiver
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
components:
  schemas:
    OpenChannelParams:
      type: object
      properties:
        senderAddress:
          type: string
        signerPubKey:
          type: string
        receiverAddress:
          type: string
        initialDeposit:
          type: integer
        expirationDate:
          type: integer
        groupId:
          type: string
    UpdateChannelParams:
      type: object
      properties:
        channelId:
          type: string
        addDeposit:
          type: integer
        expirationDate:
          type: integer
        userAddress:
          type: string
        senderAddress:
          type: string
    ClaimParams:
      type: object
      properties:
        channelId:
          type: string
        signature:
          type: string
        amount:
          type: integer
        senderAddress:
          type: string
        finalize:
          type: boolean
    CloseParams:
      type: object
      properties:
        channelId:
          type: string
        senderAddress:
          type: string
    Channel:
      type: object
      properties:
        txHash:
          type: string
        outputIndex:
          type: integer
        channelId:
          type: string
        nonce:
          type: integer
        active:
          type: boolean
        expirationDate:
          type: integer
        balance:
          type: object
        senderAddress:
          type: string
        signerAddress:
          type: string
        receiverAddress:
          type: string